refact: SNS model 패키지 분할 및 파일 분할
