+++ Tlog/app/src/main/java/com/tlog/data/model/request/sns/FollowRequest.kt
@@ -1,0 +1,6 @@
+package com.tlog.data.model.request.sns
+
+data class FollowRequest(
+    val from_userId: String,
+    val to_userId: String
+)
