+++ Tlog/app/src/main/java/com/tlog/data/model/request/sns/PostWriteBody.kt
@@ -1,0 +1,8 @@
+package com.tlog.data.model.request.sns
+
+data class PostWriteBody(
+    val author: String,
+    val courseId: String,
+    val content: String,
+    val imageUrls: List<String>
+)
