+++ Tlog/app/src/main/java/com/tlog/data/model/request/sns/ReportRequest.kt
@@ -1,0 +1,5 @@
+package com.tlog.data.model.request.sns
+
+data class ReportRequest(
+    val postId: String
+)
