+++ Tlog/app/src/main/java/com/tlog/data/model/request/sns/StatusMessage.kt
@@ -1,0 +1,6 @@
+package com.tlog.data.model.request.sns
+
+data class StatusMessage(
+    val status: Boolean,
+    val message: String
+)
