+++ Tlog/app/src/main/java/com/tlog/data/model/request/sns/UpdateSnsIdRequest.kt
@@ -1,0 +1,5 @@
+package com.tlog.data.model.request.sns
+
+data class UpdateSnsIdRequest(
+    val snsId: String
+)
