+++ Tlog/app/src/main/java/com/tlog/data/model/response/sns/ChatMessageHistory.kt
@@ -1,0 +1,11 @@
+package com.tlog.data.model.response.sns
+
+data class ChatMessageHistory(
+    val id: Long,
+    val chatRoomId: Long,
+    val senderId: String,
+    val senderName: String,
+    val content: String,
+    val sendAt: String,
+    val unreadCount: Int
+)
