+++ Tlog/app/src/main/java/com/tlog/data/model/response/sns/ChatMessageHistoryResponse.kt
@@ -1,0 +1,7 @@
+package com.tlog.data.model.response.sns
+
+data class ChatMessageHistoryResponse(
+    val messages: List<ChatMessageHistory>,
+    val nextCursor: Long?,
+    val hasNext: Boolean
+)
