+++ Tlog/app/src/main/java/com/tlog/data/model/response/sns/ChatRoom.kt
@@ -1,0 +1,9 @@
+package com.tlog.data.model.response.sns
+
+data class ChatRoom(
+    val chatRoomId: Long,
+    val lastMessageContent: String?,
+    val lastMessageSentAt: String?,
+    val countChatRoomUsers: Int,
+    val unreadCount: Int
+)
