+++ Tlog/app/src/main/java/com/tlog/data/model/response/sns/CommentRequest.kt
@@ -1,0 +1,6 @@
+package com.tlog.data.model.response.sns
+
+data class CommentRequest(
+    val author: String,
+    val content: String
+)
