+++ Tlog/app/src/main/java/com/tlog/data/model/response/sns/SnsPost.kt
@@ -1,0 +1,17 @@
+package com.tlog.data.model.response.sns
+
+import com.tlog.data.model.sns.Comment
+
+data class SnsPost(
+    val postId: String,
+    val postLikeCount: Int,
+    val postLinkCode: String,
+    val courseId: String,
+    val courseDistrics: List<String>,
+    val authorId: String,
+    val authorName: String,
+    val authorProfileImageUrl: String? = "",
+    val contentImageUrls: List<String>,
+    val content: String,
+    val replies: List<Comment>
+)
