+++ Tlog/app/src/main/java/com/tlog/data/model/response/sns/SnsPostPreview.kt
@@ -1,0 +1,6 @@
+package com.tlog.data.model.response.sns
+
+data class SnsPostPreview(
+    val postId: String,
+    val previewImageUrl: String
+)
