+++ Tlog/app/src/main/java/com/tlog/data/model/response/sns/SnsUserProfile.kt
@@ -1,0 +1,13 @@
+package com.tlog.data.model.response.sns
+
+import com.tlog.data.model.response.base.BaseListPage
+
+data class SnsUserProfile(
+    val username: String,
+    val profileImageUrl: String?,
+    val snsDescription: String?,
+    val postCount: Int,
+    val followerCount: Int,
+    val followingCount: Int,
+    val posts: BaseListPage<List<SnsPostPreview>>
+)
