+++ Tlog/app/src/main/java/com/tlog/ui/screen/sns/ChatListScreen.kt
@@ -34,7 +34,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
-import com.tlog.data.api.ChatRoom
+import com.tlog.data.model.response.sns.ChatRoom
 import com.tlog.ui.theme.MainFont
 import com.tlog.viewmodel.sns.ChatListViewModel
