+++ Tlog/app/src/main/java/com/tlog/ui/screen/sns/SnsProfileScreen.kt
@@ -30,8 +30,8 @@ import com.tlog.viewmodel.sns.SnsMyPageViewModel
 import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.TextStyle
-import com.tlog.data.api.SnsPostPreview
-import com.tlog.data.api.SnsUserProfile
+import com.tlog.data.model.response.sns.SnsPostPreview
+import com.tlog.data.model.response.sns.SnsUserProfile
 import com.tlog.ui.style.Body1Bold
 import com.tlog.ui.theme.MainFont
 import com.tlog.viewmodel.base.BaseViewModel.UiEvent
