+++ Tlog/app/src/main/java/com/tlog/viewmodel/sns/ChatListViewModel.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.tlog.api.SnsApi
-import com.tlog.data.api.ChatRoom
+import com.tlog.data.model.response.sns.ChatRoom
 import com.tlog.data.local.UserPreferences
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
