+++ Tlog/app/src/main/java/com/tlog/viewmodel/sns/SnsDetailViewModel.kt
@@ -3,7 +3,7 @@ package com.tlog.viewmodel.sns
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.State
 import com.tlog.data.local.TokenProvider
-import com.tlog.data.api.SnsPost
+import com.tlog.data.model.response.sns.SnsPost
 import com.tlog.data.local.FollowManager
 import com.tlog.data.repository.SnsRepository
 import com.tlog.ui.navigation.Screen
