+++ Tlog/app/src/main/java/com/tlog/viewmodel/sns/SnsMyPageViewModel.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import com.tlog.viewmodel.base.BaseViewModel
 import com.tlog.data.local.TokenProvider
-import com.tlog.data.api.SnsUserProfile
+import com.tlog.data.model.response.sns.SnsUserProfile
 import com.tlog.data.local.FollowManager
 import com.tlog.data.repository.SnsRepository
 import com.tlog.ui.navigation.Screen
